pyproject.toml
[project]
name = "retail-poc-video-capture"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Kevin Denver", email = "1027637+ToastShaman@users.noreply.github.com"},
]
dependencies = ["opencv-python>=4.11.0.86", "numpy>=2.2.3", "pyaudio>=0.2.14", "rerun-sdk>=0.22.1", "ruff>=0.9.10", "ultralytics>=8.3.23", "pyzmq>=26.2.1", "asyncio>=3.4.3"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
